Move BindingConfiguration and related classes to Core
