+++ src/Core/Binding/ApplicableQualityProfile.cs
