+++ src/Core/Binding/BindingConfiguration.cs
@@ -29,17 +29,12 @@ namespace SonarLint.VisualStudio.Integration.NewConnectedMode
 
         public static BindingConfiguration CreateBoundConfiguration(BoundSonarQubeProject project, SonarLintMode sonarLintMode)
         {
-            if (project == null)
-            {
-                throw new ArgumentNullException(nameof(project));
-            }
-
             return new BindingConfiguration(project, sonarLintMode);
         }
 
-        internal /*for testing purposes*/ BindingConfiguration(BoundSonarQubeProject project, SonarLintMode mode)
+        public BindingConfiguration(BoundSonarQubeProject project, SonarLintMode mode)
         {
-            this.Project = project;
+            this.Project = project ?? throw new ArgumentNullException(nameof(project));
             this.Mode = mode;
         }
