+++ src/Core/Binding/BoundSonarQubeProjectExtensions.cs
@@ -23,7 +23,7 @@ using SonarQube.Client.Models;
 
 namespace SonarLint.VisualStudio.Integration.Persistence
 {
-    internal static class BoundSonarQubeProjectExtensions
+    public static class BoundSonarQubeProjectExtensions
     {
         public static ConnectionInformation CreateConnectionInformation(this BoundSonarQubeProject binding)
         {
